feat: ResponseEntity<?> addUser
The new User object is saved as before. But the
resulting object is wrapped using the UserModelAssembler.
Spring MVC’s ResponseEntity is used to create an
HTTP 201 Created status message. This type of response
typically includes a Location response header, and we use
the URI derived from the model’s self-related link.
Additionally, return the model-based version of the saved object.